Refactoring code smell. 1. ProductDal named EfProductDal. 2. IProductDal created. 3. Form1 updated.
