+++ Northwind.Business/Concrete/ProductManager.cs
@@ -1,4 +1,5 @@
-﻿using Nothwind.DataAccess.Concrete;
+﻿using Nothwind.DataAccess.Abstract;
+using Nothwind.DataAccess.Concrete.EntityFramework;
 using Nothwind.Entities.Concrete;
 using System;
 using System.Collections.Generic;
@@ -10,7 +11,13 @@ namespace Northwind.Business.Concrete
 {
     public class ProductManager
     {
-        ProductDal _productDal = new ProductDal();
+        private IProductDal _productDal;
+
+        public ProductManager(IProductDal productDal)
+        {
+            _productDal = productDal;
+        }
+
         public List<Product> GetAll()
         {
             return _productDal.GetAll();
