+++ Nothwind.DataAccess/Abstract/IProductDal.cs
@@ -1,0 +1,18 @@
+﻿using Nothwind.Entities.Concrete;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Nothwind.DataAccess.Abstract
+{
+    public interface IProductDal
+    {
+        List<Product> GetAll();
+        Product Get(int id);
+        void Add(Product product);
+        void Update(Product product);
+        void Delete(Product product);
+    }
+}
