+++ Nothwind.DataAccess/Concrete/EntityFramework/NorthwindContext.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Nothwind.DataAccess.Concrete
+namespace Nothwind.DataAccess.Concrete.EntityFramework
 {
     public class NorthwindContext : DbContext
     {
